requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
python-dotenv==1.0.1
django-jazzmin
